Separate code into util file
